Fixing output to console when no tty is present
Had to remove blessings since it was too smart and wasn't happy when the
container was run without the `-t` flag. Instead, the ANSI escape codes are
hardcoded like a barbaric programmer.

Changed the way the script is invoked to allow for the python `-u` flag which
gives unbuffered stdout and stderr.

Touched up the `GET /api/alive` message a little as well.