Using make to build the telemetry service image as well
This makes it a lot easier to automate cross-compiled builds later as it's not
necessary to build the autogenerated protobuf message files with a cross-
compiler.

Added a `.gitignore` specifically for the service since protobuf messages files
will be created in the `telemetry/` directory from `make proto`.

The root level `Makefile` was added to automate building across all services.
It also takes cares of building any dependencies in `common/` before making
other services.