Adding the initial version of the interop Protobufs
Note that the dependency exprotobuf cannot turn recursively turn maps
into the message structs, so a function `Message.form_message/2` was
created to do that.

For the Interop Proxy service to compile, it expects a `lib/messages`
directory with the message definitions in `common/messages`. When a
Dockerfile is made it'll take of moving that in at compile-time.