Adding `POST /api/telemetry`
To get this going we also needed to be able to parse the body (whether
it be Protobuf or JSON). A custom Plug was added for this to work.

I manually tested this with JSON and it appears to work, haven't tested
the Protobuf side yet.

A message was added for generic replies from the interop server, in case
some service wanted to see the message.