Handling Protobuf to JSON more correctly.
According to the Protobuf spec, binary data is supposed to be encoded in
base64 in JSON, and enums are supposed to go by their string names. I
reflected these two things in the encoding and decoding process.

The image_base64 field has been removed from the Odlcs since it's not
needed anymore.

Added JSON tests for the OdlcController.