Initial file watching implementation.
Created a backend for watching for new files in the /opt/new-images
directory. This is mostly useful for testing (be it manually or in unit
tests).

Added a basic image storage system which keeps all data (including image
metadata) in files. Images are to be stored as PNGs, and the metadata is
simply a serialized `imagery.Image` protobuf message. The metadata for
the file watching backend only has the time that the image was found.
The time is not going to accurate enough to pair telemetry to it.

Made the API combine the metadata with the image and serving it over
`GET /api/image/:id`.