Making returning the image or warped image optional
By default we'll return the original image, and if requested, we'll
return the warped one as well (and if we have it).

This is just set by query params on the request... will document that in
the README later.

Did a little refactor to make this work, and made the two image
endpoints a little cleaner.