Initial sync backend implementation
For this backend, a loop runs each 100 ms and checks to see what the
latest image id number is, then it fetches all the images it doesn't
have.

If the service that is being synced to crashes, it's id numbers will
reset and cause the syncing service to crash for now.