Creating .rs message files with make, updating ignore files
To get this going I added `--rust_out` to the protoc image so we can use
it here for rust generation.

Deleted previous message files and put in some that can be generated
inside a messages folder. Also updated the references in `main.js`.

Updated the dockerignore and gitignore files to ignore the Cargo.lock
file since this project's dependencies are going to be a bit volatile.