Adding timeout to request, and using async/forever
Requests will now timeout after a set amount of time (by default 5
seconds). If the request times out, then we say the service is offline.
Instead of using `setInterval(...)` now the next request will start 3
seconds after the previous by using async/forever. This cuts down some
of the logic that was being used before if if the next interval happened
while a request was already going on.