Add imagery endpoint for getting the list of available images
The existing method is to check the imagery count for how many images
there are to figure out what the ids are. However, this causes problems
for adding in image retention policies, or syncing images in parallel
where an image with a higher id number might be available first.

In the meantime, before a change is made, the available endpoint just
lists ids 0 to count - 1, to allow for other services to be able to
continue getting the list of images in a more future-proof way.