Disable source-map-support during tests to fix test error reporting
It seems like there was a conflict with the built-in source mapping
somewhere in Jest and the source-map-support babel plugin. It was
working fine with the service source code, but not for the test cases.
The `.babelrc`s are now configured to only use the source-map-support
plugin when building the source code, and is not including when running
tests, since Jest takes care of this.