Start asyncio rewrite of image-rec-master
This starts a series of commits turning image-rec-master into an
asyncio-based Python service. This allows for easier development since
this service pairs with image-rec-worker.

The planned design for image-rec-master is to be stateless, so that all
state is stored in Redis. Currently, all the planned pipeline lists and
sets are currently being returned in an endpoint. Background tasks which
move images into these lists and sets will be added in future commits.

The pipeline can also be reset if need be.

Both sets and lists are being used for some image and target ids. Lists
are used when order matters, or if it is advantageous to use a Redis
command like `BRPOPLPUSH`. Unprocessed and processing items are
considered to be lists because of this. However, the pipeline state
endpoint will return all the data ordered regardless if it is a list or
set.