Ensure watch and multi/exec are on the same connection
An async context manager was added to acquire a connection and then wrap
that into a new Redis client object, to make sure that a watch and
multi/exec operation does not get split across multiple connections.

Since `@contextlib/asynccontextmanager` is available in Python 3.7, the
Python version was updated as well.