Use a mirror of the interop server for testing
The images on the `auvsisuas` Docker Hub organization are being
deleted, and because of how fast the old one was removed given that the
new one had been out a little more than a week, a mirror is needed to
prevent build failures and other development problems.

In the future this also allows for mirroring images in the middle of a
month, instead of waiting until the next.