z-cam-e1: update the tests to match
We've now got an (optional) param in the z-cam backend that will skip the configuration ritual, mostly for the sake of testing (I don't really see much sense in mocking all of the endpoints we hit while configuring the camera).
This param doesn't have a matching env var; I don't think this is something that we'll actually use outside of for tests.