z-cam-e1: more tweaks from before competition
This was the only configuration that actually worked on the camera.

The options that are commented out aren't strictly necessary and we ditched them in the interest of time. If we want to be thorough, someone should check that they work as expected with the camera.

Also, there's some subtle/undocumented behavior for some of the options:
  - sharpness can't actually be set (see: ' Normal ')
  - we set lcd to On, but then set the timeout so that:
    + the screen is off for the flight
    + but, humans can still turn the screen back on without needing to write `curl` commands
  - more in the comments